beta_backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "catch-your-partner-backend"
version = "0.4.0"
description = "Modular FastAPI backend for Catch Your Partner core flow"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.110.0",
  "uvicorn[standard]==0.29.0",
  "pydantic==2.6.4",
  "pydantic-settings==2.2.1",
  "email-validator==2.1.1",
  "SQLAlchemy==2.0.29",
  "alembic==1.13.1",
  "psycopg[binary]==3.1.18",
  "PyJWT==2.8.0",
  "passlib[bcrypt]==1.7.4",
  "httpx==0.27.0",
  "pytest==8.1.1",
  "pytest-cov==5.0.0",
  "pytest-asyncio==0.23.6",
  "python-multipart==0.0.9",
  "Pillow==10.3.0",
  "numpy==1.26.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
filterwarnings = [
  "ignore:The 'app' shortcut is now deprecated:DeprecationWarning:httpx._client",
]
